Cleaned up npm related references, split tests, .NET Core blazor related detection
