+++ src/Detector/DotNetCore/DotnetCoreConstants.cs
@@ -27,8 +27,6 @@ namespace Microsoft.Oryx.Detector.DotNetCore
 
         public const string AzureFunctionsPackageReference = "Microsoft.NET.Sdk.Functions";
 
-        public const string AzureBlazorWasmRazorLangVersionXPathExpression =
-            "/Project/PropertyGroup/RazorLangVersion";
         public const string AzureBlazorWasmPackageReference = "Microsoft.AspNetCore.Components.WebAssembly";
     }
 }
