+++ tests/Oryx.BuildImage.Tests/Python/CondaTests.cs
@@ -1,0 +1,241 @@
+﻿// --------------------------------------------------------------------------------------------
+// Copyright (c) Microsoft Corporation. All rights reserved.
+// Licensed under the MIT license.
+// --------------------------------------------------------------------------------------------
+
+using System.Collections.Generic;
+using System.IO;
+using Microsoft.Oryx.BuildScriptGenerator.Common;
+using Microsoft.Oryx.Tests.Common;
+using Xunit;
+using Xunit.Abstractions;
+
+namespace Microsoft.Oryx.BuildImage.Tests
+{
+    public class CondaTests : PythonSampleAppsTestBase
+    {
+        public DockerVolume CreateCondaSampleAppVolume(string sampleAppName) =>
+            DockerVolume.CreateMirror(Path.Combine(_hostSamplesDir, "python", "conda-samples", sampleAppName));
+
+        public CondaTests(ITestOutputHelper output) : base(output)
+        {
+        }
+
+        [Fact]
+        public void CanBuildAppWhichHasRequirementsTxtFile()
+        {
+            // Arrange
+            var appName = "requirements";
+            var volume = CreateCondaSampleAppVolume(appName);
+            var appDir = volume.ContainerDir;
+            var appOutputDir = "/tmp/app-output";
+            var script = new ShellScriptBuilder()
+                .AddBuildCommand($"{appDir} -o {appOutputDir}")
+                .AddDirectoryExistsCheck($"{appOutputDir}/venv")
+                // Following command makes sure that this package 'matplotlib' is present
+                .AddCommand("source /opt/conda/etc/profile.d/conda.sh")
+                .AddCommand($"conda activate {appOutputDir}/venv")
+                .AddCommand("pip show matplotlib")
+                .ToString();
+
+            // Act
+            var result = _dockerCli.Run(new DockerRunArguments
+            {
+                ImageId = _imageHelper.GetVsoBuildImage(),
+                EnvironmentVariables = new List<EnvironmentVariable> { CreateAppNameEnvVar(appName) },
+                Volumes = new List<DockerVolume> { volume },
+                CommandToExecuteOnRun = "/bin/bash",
+                CommandArguments = new[] { "-c", script }
+            });
+
+            // Assert
+            RunAsserts(
+                () =>
+                {
+                    Assert.True(result.IsSuccess);
+                },
+                result.GetDebugInfo());
+        }
+
+        [Fact]
+        public void CanBuildAppWhichHasCondaEnvironmentYmlFile()
+        {
+            // Arrange
+            var appName = "conda";
+            var volume = CreateCondaSampleAppVolume(appName);
+            var appDir = volume.ContainerDir;
+            var appOutputDir = "/tmp/app-output";
+            var script = new ShellScriptBuilder()
+                .AddBuildCommand($"{appDir} -o {appOutputDir}")
+                .AddDirectoryExistsCheck($"{appOutputDir}/venv")
+                // Following command makes sure that this package 'matplotlib' is present
+                .AddCommand("source /opt/conda/etc/profile.d/conda.sh")
+                .AddCommand($"conda activate {appOutputDir}/venv")
+                .AddCommand("pip show matplotlib")
+                .ToString();
+
+            // Act
+            var result = _dockerCli.Run(new DockerRunArguments
+            {
+                ImageId = _imageHelper.GetVsoBuildImage(),
+                EnvironmentVariables = new List<EnvironmentVariable> { CreateAppNameEnvVar(appName) },
+                Volumes = new List<DockerVolume> { volume },
+                CommandToExecuteOnRun = "/bin/bash",
+                CommandArguments = new[] { "-c", script }
+            });
+
+            // Assert
+            RunAsserts(
+                () =>
+                {
+                    Assert.True(result.IsSuccess);
+                },
+                result.GetDebugInfo());
+        }
+
+        [Fact]
+        public void CanBuildPython2AppHavingRequirementsTxtFile()
+        {
+            // Arrange
+            var appName = "python2_runtime";
+            var volume = CreateCondaSampleAppVolume(appName);
+            var appDir = volume.ContainerDir;
+            var appOutputDir = "/tmp/app-output";
+            var script = new ShellScriptBuilder()
+                .AddBuildCommand($"{appDir} -o {appOutputDir}")
+                .AddDirectoryExistsCheck($"{appOutputDir}/venv")
+                // Following command makes sure that this package 'matplotlib' is present
+                .AddCommand("source /opt/conda/etc/profile.d/conda.sh")
+                .AddCommand($"conda activate {appOutputDir}/venv")
+                .AddCommand("pip show matplotlib")
+                .ToString();
+
+            // Act
+            var result = _dockerCli.Run(new DockerRunArguments
+            {
+                ImageId = _imageHelper.GetVsoBuildImage(),
+                EnvironmentVariables = new List<EnvironmentVariable> { CreateAppNameEnvVar(appName) },
+                Volumes = new List<DockerVolume> { volume },
+                CommandToExecuteOnRun = "/bin/bash",
+                CommandArguments = new[] { "-c", script }
+            });
+
+            // Assert
+            RunAsserts(
+                () =>
+                {
+                    Assert.True(result.IsSuccess);
+                },
+                result.GetDebugInfo());
+        }
+
+        [Fact]
+        public void CanBuildAppWithCondaEnviornmentYmlFileHavingPipPackages()
+        {
+            // Arrange
+            var appName = "python-conda_pip";
+            var volume = CreateCondaSampleAppVolume(appName);
+            var appDir = volume.ContainerDir;
+            var appOutputDir = "/tmp/app-output";
+            var script = new ShellScriptBuilder()
+                .AddBuildCommand($"{appDir} -o {appOutputDir}")
+                .AddDirectoryExistsCheck($"{appOutputDir}/venv")
+                // Following command makes sure that this package 'matplotlib' is present
+                .AddCommand("source /opt/conda/etc/profile.d/conda.sh")
+                .AddCommand($"conda activate {appOutputDir}/venv")
+                .AddCommand("pip show matplotlib")
+                .ToString();
+
+            // Act
+            var result = _dockerCli.Run(new DockerRunArguments
+            {
+                ImageId = _imageHelper.GetVsoBuildImage(),
+                EnvironmentVariables = new List<EnvironmentVariable> { CreateAppNameEnvVar(appName) },
+                Volumes = new List<DockerVolume> { volume },
+                CommandToExecuteOnRun = "/bin/bash",
+                CommandArguments = new[] { "-c", script }
+            });
+
+            // Assert
+            RunAsserts(
+                () =>
+                {
+                    Assert.True(result.IsSuccess);
+                },
+                result.GetDebugInfo());
+        }
+
+        [Fact]
+        public void CanBuildJuliaPythonSampleApp()
+        {
+            // Arrange
+            var appName = "julia-python";
+            var volume = CreateCondaSampleAppVolume(appName);
+            var appDir = volume.ContainerDir;
+            var appOutputDir = "/tmp/app-output";
+            var script = new ShellScriptBuilder()
+                .AddBuildCommand($"{appDir} -o {appOutputDir}")
+                .AddDirectoryExistsCheck($"{appOutputDir}/venv")
+                // Following command makes sure that this package 'matplotlib' is present
+                .AddCommand("source /opt/conda/etc/profile.d/conda.sh")
+                .AddCommand($"conda activate {appOutputDir}/venv")
+                .AddCommand("pip show julia")
+                .ToString();
+
+            // Act
+            var result = _dockerCli.Run(new DockerRunArguments
+            {
+                ImageId = _imageHelper.GetVsoBuildImage(),
+                EnvironmentVariables = new List<EnvironmentVariable> { CreateAppNameEnvVar(appName) },
+                Volumes = new List<DockerVolume> { volume },
+                CommandToExecuteOnRun = "/bin/bash",
+                CommandArguments = new[] { "-c", script }
+            });
+
+            // Assert
+            RunAsserts(
+                () =>
+                {
+                    Assert.True(result.IsSuccess);
+                },
+                result.GetDebugInfo());
+        }
+
+        [Fact]
+        public void CanBuildJupiterRiseApp()
+        {
+            // Arrange
+            var appName = "jupyter-rise";
+            var volume = CreateCondaSampleAppVolume(appName);
+            var appDir = volume.ContainerDir;
+            var appOutputDir = "/tmp/app-output";
+            var script = new ShellScriptBuilder()
+                .AddBuildCommand($"{appDir} -o {appOutputDir}")
+                .AddDirectoryExistsCheck($"{appOutputDir}/venv")
+                // Following command makes sure that this package 'matplotlib' is present
+                .AddCommand("source /opt/conda/etc/profile.d/conda.sh")
+                .AddCommand($"conda activate {appOutputDir}/venv")
+                .AddCommand("pip show matplotlib")
+                .ToString();
+
+            // Act
+            var result = _dockerCli.Run(new DockerRunArguments
+            {
+                ImageId = _imageHelper.GetVsoBuildImage(),
+                EnvironmentVariables = new List<EnvironmentVariable> { CreateAppNameEnvVar(appName) },
+                Volumes = new List<DockerVolume> { volume },
+                CommandToExecuteOnRun = "/bin/bash",
+                CommandArguments = new[] { "-c", script }
+            });
+
+            // Assert
+            RunAsserts(
+                () =>
+                {
+                    Assert.True(result.IsSuccess);
+                },
+                result.GetDebugInfo());
+        }
+    }
+
+}
