+++ tests/Oryx.BuildImage.Tests/Python/PythonSampleAppsTest.cs
@@ -16,23 +16,11 @@ using Xunit.Abstractions;
 
 namespace Microsoft.Oryx.BuildImage.Tests
 {
-    public class PythonSampleAppsTestBase : SampleAppsTestBase
-    {
-        public const string PackagesDirectory = "__oryx_packages__";
-
-        public DockerVolume CreateSampleAppVolume(string sampleAppName) =>
-            DockerVolume.CreateMirror(Path.Combine(_hostSamplesDir, "python", sampleAppName));
-
-        public PythonSampleAppsTestBase(ITestOutputHelper output) : base(output)
-        {
-        }
-    }
-
-    public class PythonSampleAppsOtherTests : PythonSampleAppsTestBase, IClassFixture<TestTempDirTestFixture>
+    public class PythonSampleAppsTest : PythonSampleAppsTestBase, IClassFixture<TestTempDirTestFixture>
     {
         protected readonly string _tempDirRootPath;
 
-        public PythonSampleAppsOtherTests(ITestOutputHelper output, TestTempDirTestFixture testFixture)
+        public PythonSampleAppsTest(ITestOutputHelper output, TestTempDirTestFixture testFixture)
             : base(output)
         {
             _tempDirRootPath = testFixture.RootDirPath;
@@ -1219,367 +1207,6 @@ namespace Microsoft.Oryx.BuildImage.Tests
         {
             var ver = new SemVer.Version(version);
             return $"pythonenv{ver.Major}.{ver.Minor}";
-        }
-    }
-
-    public class PythonSampleAppsDjangoAppRunningCollectStatic : PythonSampleAppsTestBase
-    {
-        public PythonSampleAppsDjangoAppRunningCollectStatic(ITestOutputHelper output) : base(output)
-        {
-        }
-
-        [Theory]
-        [InlineData(null)]
-        [InlineData("false")]
-        public void GeneratesScript_AndBuilds_DjangoApp_RunningCollectStatic(string disableCollectStatic)
-        {
-            // Arrange
-            var appName = "django-app";
-            var volume = CreateSampleAppVolume(appName);
-            var appDir = volume.ContainerDir;
-            var appOutputDir = "/tmp/app-output";
-            var scriptBuilder = new ShellScriptBuilder();
-            if (string.IsNullOrEmpty(disableCollectStatic))
-            {
-                scriptBuilder.AddCommand(
-                    $"export {EnvironmentSettingsKeys.DisableCollectStatic}={disableCollectStatic}");
-            }
-            var script = scriptBuilder
-                .AddBuildCommand(
-                $"{appDir} -o {appOutputDir} --platform {PythonConstants.PlatformName} --platform-version {PythonVersions.Python37Version}")
-                // These css files should be available since 'collectstatic' is run in the script
-                .AddFileExistsCheck($"{appOutputDir}/staticfiles/css/boards.css")
-                .AddFileExistsCheck($"{appOutputDir}/staticfiles/css/uservoice.css")
-                .ToString();
-
-            // Act
-            var result = _dockerCli.Run(new DockerRunArguments
-            {
-                ImageId = Settings.BuildImageName,
-                EnvironmentVariables = new List<EnvironmentVariable> { CreateAppNameEnvVar(appName) },
-                Volumes = new List<DockerVolume> { volume },
-                CommandToExecuteOnRun = "/bin/bash",
-                CommandArguments = new[] { "-c", script }
-            });
-
-            // Assert
-            RunAsserts(
-                () =>
-                {
-                    Assert.True(result.IsSuccess);
-                },
-                result.GetDebugInfo());
-        }
-    }
-
-    public class PythonSampleAppsBuildsShapely : PythonSampleAppsTestBase
-    {
-        public PythonSampleAppsBuildsShapely(ITestOutputHelper output) : base(output)
-        {
-        }
-
-        [Theory]
-        [MemberData(nameof(TestValueGenerator.GetPythonVersions), MemberType = typeof(TestValueGenerator))]
-        public void GeneratesScript_AndBuilds_Shapely_With_Python(string version)
-        {
-            // Arrange
-            var appName = "shapely-flask-app";
-            var volume = CreateSampleAppVolume(appName);
-            var appDir = volume.ContainerDir;
-            var appOutputDir = "/tmp/app-output";
-            var script = new ShellScriptBuilder()
-                .AddBuildCommand($"{appDir} -o {appOutputDir} --platform {PythonConstants.PlatformName} --platform-version {version}")
-                .ToString();
-
-            // Act
-            var result = _dockerCli.Run(new DockerRunArguments
-            {
-                ImageId = Settings.BuildImageName,
-                EnvironmentVariables = new List<EnvironmentVariable> { CreateAppNameEnvVar(appName) },
-                Volumes = new List<DockerVolume> { volume },
-                CommandToExecuteOnRun = "/bin/bash",
-                CommandArguments = new[] { "-c", script }
-            });
-
-            // Assert
-            RunAsserts(
-                () =>
-                {
-                    Assert.True(result.IsSuccess);
-                },
-                result.GetDebugInfo());
-        }
-    }
-
-    public class PythonSampleAppsDjangoAppWithoutRunningCollectStatic : PythonSampleAppsTestBase
-    {
-        public PythonSampleAppsDjangoAppWithoutRunningCollectStatic(ITestOutputHelper output) : base(output)
-        {
-        }
-
-        [Theory]
-        [InlineData("true")]
-        [InlineData("True")]
-        public void GeneratesScript_AndBuilds_DjangoApp_WithoutRunningCollectStatic_IfDisableCollectStatic_IsTrue(
-            string disableCollectStatic)
-        {
-            // Arrange
-            var appName = "django-app";
-            var volume = CreateSampleAppVolume(appName);
-            var appDir = volume.ContainerDir;
-            var appOutputDir = "/tmp/app-output";
-            var script = new ShellScriptBuilder()
-                .AddCommand($"export {EnvironmentSettingsKeys.DisableCollectStatic}={disableCollectStatic}")
-                .AddBuildCommand(
-                $"{appDir} -o {appOutputDir} --platform {PythonConstants.PlatformName} --platform-version {PythonVersions.Python37Version}")
-                // These css files should NOT be available since 'collectstatic' is set off
-                .AddFileDoesNotExistCheck($"{appOutputDir}/staticfiles/css/boards.css")
-                .AddFileDoesNotExistCheck($"{appOutputDir}/staticfiles/css/uservoice.css")
-                .ToString();
-
-            // Act
-            var result = _dockerCli.Run(new DockerRunArguments
-            {
-                ImageId = Settings.BuildImageName,
-                EnvironmentVariables = new List<EnvironmentVariable> { CreateAppNameEnvVar(appName) },
-                Volumes = new List<DockerVolume> { volume },
-                CommandToExecuteOnRun = "/bin/bash",
-                CommandArguments = new[] { "-c", script }
-            });
-
-            // Assert
-            RunAsserts(
-                () =>
-                {
-                    Assert.True(result.IsSuccess);
-                },
-                result.GetDebugInfo());
-        }
-    }
-
-    public class CondaTests : PythonSampleAppsTestBase
-    {
-        public DockerVolume CreateCondaSampleAppVolume(string sampleAppName) =>
-            DockerVolume.CreateMirror(Path.Combine(_hostSamplesDir, "python", "conda-samples", sampleAppName));
-
-        public CondaTests(ITestOutputHelper output) : base(output)
-        {
-        }
-
-        [Fact]
-        public void CanBuildAppWhichHasRequirementsTxtFile()
-        {
-            // Arrange
-            var appName = "requirements";
-            var volume = CreateCondaSampleAppVolume(appName);
-            var appDir = volume.ContainerDir;
-            var appOutputDir = "/tmp/app-output";
-            var script = new ShellScriptBuilder()
-                .AddBuildCommand($"{appDir} -o {appOutputDir}")
-                .AddDirectoryExistsCheck($"{appOutputDir}/venv")
-                // Following command makes sure that this package 'matplotlib' is present
-                .AddCommand("source /opt/conda/etc/profile.d/conda.sh")
-                .AddCommand($"conda activate {appOutputDir}/venv")
-                .AddCommand("pip show matplotlib")
-                .ToString();
-
-            // Act
-            var result = _dockerCli.Run(new DockerRunArguments
-            {
-                ImageId = _imageHelper.GetVsoBuildImage(),
-                EnvironmentVariables = new List<EnvironmentVariable> { CreateAppNameEnvVar(appName) },
-                Volumes = new List<DockerVolume> { volume },
-                CommandToExecuteOnRun = "/bin/bash",
-                CommandArguments = new[] { "-c", script }
-            });
-
-            // Assert
-            RunAsserts(
-                () =>
-                {
-                    Assert.True(result.IsSuccess);
-                },
-                result.GetDebugInfo());
-        }
-
-        [Fact]
-        public void CanBuildAppWhichHasCondaEnvironmentYmlFile()
-        {
-            // Arrange
-            var appName = "conda";
-            var volume = CreateCondaSampleAppVolume(appName);
-            var appDir = volume.ContainerDir;
-            var appOutputDir = "/tmp/app-output";
-            var script = new ShellScriptBuilder()
-                .AddBuildCommand($"{appDir} -o {appOutputDir}")
-                .AddDirectoryExistsCheck($"{appOutputDir}/venv")
-                // Following command makes sure that this package 'matplotlib' is present
-                .AddCommand("source /opt/conda/etc/profile.d/conda.sh")
-                .AddCommand($"conda activate {appOutputDir}/venv")
-                .AddCommand("pip show matplotlib")
-                .ToString();
-
-            // Act
-            var result = _dockerCli.Run(new DockerRunArguments
-            {
-                ImageId = _imageHelper.GetVsoBuildImage(),
-                EnvironmentVariables = new List<EnvironmentVariable> { CreateAppNameEnvVar(appName) },
-                Volumes = new List<DockerVolume> { volume },
-                CommandToExecuteOnRun = "/bin/bash",
-                CommandArguments = new[] { "-c", script }
-            });
-
-            // Assert
-            RunAsserts(
-                () =>
-                {
-                    Assert.True(result.IsSuccess);
-                },
-                result.GetDebugInfo());
-        }
-
-        [Fact]
-        public void CanBuildPython2AppHavingRequirementsTxtFile()
-        {
-            // Arrange
-            var appName = "python2_runtime";
-            var volume = CreateCondaSampleAppVolume(appName);
-            var appDir = volume.ContainerDir;
-            var appOutputDir = "/tmp/app-output";
-            var script = new ShellScriptBuilder()
-                .AddBuildCommand($"{appDir} -o {appOutputDir}")
-                .AddDirectoryExistsCheck($"{appOutputDir}/venv")
-                // Following command makes sure that this package 'matplotlib' is present
-                .AddCommand("source /opt/conda/etc/profile.d/conda.sh")
-                .AddCommand($"conda activate {appOutputDir}/venv")
-                .AddCommand("pip show matplotlib")
-                .ToString();
-
-            // Act
-            var result = _dockerCli.Run(new DockerRunArguments
-            {
-                ImageId = _imageHelper.GetVsoBuildImage(),
-                EnvironmentVariables = new List<EnvironmentVariable> { CreateAppNameEnvVar(appName) },
-                Volumes = new List<DockerVolume> { volume },
-                CommandToExecuteOnRun = "/bin/bash",
-                CommandArguments = new[] { "-c", script }
-            });
-
-            // Assert
-            RunAsserts(
-                () =>
-                {
-                    Assert.True(result.IsSuccess);
-                },
-                result.GetDebugInfo());
-        }
-
-        [Fact]
-        public void CanBuildAppWithCondaEnviornmentYmlFileHavingPipPackages()
-        {
-            // Arrange
-            var appName = "python-conda_pip";
-            var volume = CreateCondaSampleAppVolume(appName);
-            var appDir = volume.ContainerDir;
-            var appOutputDir = "/tmp/app-output";
-            var script = new ShellScriptBuilder()
-                .AddBuildCommand($"{appDir} -o {appOutputDir}")
-                .AddDirectoryExistsCheck($"{appOutputDir}/venv")
-                // Following command makes sure that this package 'matplotlib' is present
-                .AddCommand("source /opt/conda/etc/profile.d/conda.sh")
-                .AddCommand($"conda activate {appOutputDir}/venv")
-                .AddCommand("pip show matplotlib")
-                .ToString();
-
-            // Act
-            var result = _dockerCli.Run(new DockerRunArguments
-            {
-                ImageId = _imageHelper.GetVsoBuildImage(),
-                EnvironmentVariables = new List<EnvironmentVariable> { CreateAppNameEnvVar(appName) },
-                Volumes = new List<DockerVolume> { volume },
-                CommandToExecuteOnRun = "/bin/bash",
-                CommandArguments = new[] { "-c", script }
-            });
-
-            // Assert
-            RunAsserts(
-                () =>
-                {
-                    Assert.True(result.IsSuccess);
-                },
-                result.GetDebugInfo());
-        }
-
-        [Fact]
-        public void CanBuildJuliaPythonSampleApp()
-        {
-            // Arrange
-            var appName = "julia-python";
-            var volume = CreateCondaSampleAppVolume(appName);
-            var appDir = volume.ContainerDir;
-            var appOutputDir = "/tmp/app-output";
-            var script = new ShellScriptBuilder()
-                .AddBuildCommand($"{appDir} -o {appOutputDir}")
-                .AddDirectoryExistsCheck($"{appOutputDir}/venv")
-                // Following command makes sure that this package 'matplotlib' is present
-                .AddCommand("source /opt/conda/etc/profile.d/conda.sh")
-                .AddCommand($"conda activate {appOutputDir}/venv")
-                .AddCommand("pip show julia")
-                .ToString();
-
-            // Act
-            var result = _dockerCli.Run(new DockerRunArguments
-            {
-                ImageId = _imageHelper.GetVsoBuildImage(),
-                EnvironmentVariables = new List<EnvironmentVariable> { CreateAppNameEnvVar(appName) },
-                Volumes = new List<DockerVolume> { volume },
-                CommandToExecuteOnRun = "/bin/bash",
-                CommandArguments = new[] { "-c", script }
-            });
-
-            // Assert
-            RunAsserts(
-                () =>
-                {
-                    Assert.True(result.IsSuccess);
-                },
-                result.GetDebugInfo());
-        }
-
-        [Fact]
-        public void CanBuildJupiterRiseApp()
-        {
-            // Arrange
-            var appName = "jupyter-rise";
-            var volume = CreateCondaSampleAppVolume(appName);
-            var appDir = volume.ContainerDir;
-            var appOutputDir = "/tmp/app-output";
-            var script = new ShellScriptBuilder()
-                .AddBuildCommand($"{appDir} -o {appOutputDir}")
-                .AddDirectoryExistsCheck($"{appOutputDir}/venv")
-                // Following command makes sure that this package 'matplotlib' is present
-                .AddCommand("source /opt/conda/etc/profile.d/conda.sh")
-                .AddCommand($"conda activate {appOutputDir}/venv")
-                .AddCommand("pip show matplotlib")
-                .ToString();
-
-            // Act
-            var result = _dockerCli.Run(new DockerRunArguments
-            {
-                ImageId = _imageHelper.GetVsoBuildImage(),
-                EnvironmentVariables = new List<EnvironmentVariable> { CreateAppNameEnvVar(appName) },
-                Volumes = new List<DockerVolume> { volume },
-                CommandToExecuteOnRun = "/bin/bash",
-                CommandArguments = new[] { "-c", script }
-            });
-
-            // Assert
-            RunAsserts(
-                () =>
-                {
-                    Assert.True(result.IsSuccess);
-                },
-                result.GetDebugInfo());
         }
     }
 }
